requirements.txt
python-dotenv~=1.0.0
aiohttp~=3.8.5
discord.py==2.3.2
pandas~=2.1.3
requests~=2.31.0
sqlalchemy~=2.0.41